+++ test/timezone_secs_since_dow.c
@@ -29,7 +29,7 @@ int main(void)
 		if (dow < 0) dow += 7;
 
 		// Get the result
-		result = timezone_secs_since_dow(timezonename, hour, minute, second, dow);
+		result = (int)timezone_secs_since_dow(timezonename, hour, minute, second, dow);
 
 		// Result should be within 5 seconds
 		TEST_TRUE(abs(result - 86400 * i) < 5);
